requirements.txt
# Core dependencies
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
seaborn>=0.12.0
scikit-learn>=1.3.0
joblib>=1.3.0

# HTTP requests
requests>=2.31.0

# Security (for production)
cryptography>=41.0.0

# Development
python-dotenv>=1.0.0
